Move screens components to pages folder
