+++ src/pages/WatchLater.jsx
@@ -1,13 +1,12 @@
 import { useSelector, useDispatch } from 'react-redux';
 import { Link } from 'react-router-dom';
 import watchLaterSlice from '../data/watchLaterSlice';
-import Movie from './Movie';
+import Movie from '../components/Movie';
 import '../styles/starred.scss';
 
 function WatchLater({ viewTrailer }) {
-  const state = useSelector((state) => state);
-  const { watchLater } = state;
-  const { remveAllWatchLater } = watchLaterSlice.actions;
+  const { watchLater } = useSelector((state) => state);
+  const { removeAllWatchLater } = watchLaterSlice.actions;
   const dispatch = useDispatch();
 
   return (
@@ -22,7 +21,11 @@ function WatchLater({ viewTrailer }) {
           </div>
 
           <footer className='text-center'>
-            <button className='btn btn-primary' onClick={() => dispatch(remveAllWatchLater())}>
+            <button
+              type='button'
+              className='btn btn-primary'
+              onClick={() => dispatch(removeAllWatchLater())}
+            >
               Empty list
             </button>
           </footer>
